+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] - 19 December 2020.
 
-* TODO: Describe initial release.
+* increased the Country properties to 26 
+* added 6 new models for easy data access
+* refactored the main Countries class 
+* fixed minor bugs .
+* updated readme with full instructions.
+
+## [0.0.1] - 17 December 2020.
+
+* initial release , get all countries or filter by name and 15 other properties 
+* added Country and CountryList Models